Refactor metrics endpoint
